Upd: parser: Overhaul casting behavior
The `str` and `any` types are not considered as references anymore.

For the former, this is motivated by the fact that strings are actually
arrays, and that arrays are not allowed to be treated as references in
Reflect. This was a bad behavior inherited from the C language.

For the later, this is not really motivated, this made the change easier
to do. This may prove a good or bad behavior in the future, but we will
see once the `any` type behavior will be well defined, it is still
unclear for now.

By the way, the Symbol.auto_cast function have been simplified.