Fix: parser: Allow full expressions as function parameter
Only values were allowed. This disallowed to pass operations as
paramater, for example `malloc(sizeof(int) * 8)` was not working.

By the way, the hello_world test have been renamed to function_call.
This was what it was doing, but was ways too simplistic for the current
state of the project.