Rem: parser: lvalue cast
This ambiguous as we can't detect whether we are casting to a reference
or referencing a cast. While there is still the problem for rvalue
casts, this does not cause a problem due to auto-referencing.

Another reason is that in the future we will have more complex casts,
using functions for it. In this case, the return value of a function
will not be usable as an lvalue.

In short, lvalue casts bring more problems than it solves.